feat: refactor to resolve circular depedencies
