+++ app/app.py
@@ -12,11 +12,11 @@ from fastapi.templating import Jinja2Templates
 from fastapi_utils.tasks import repeat_every
 
 from app.backend import PrecisBackend
-from app.context import GlobalSettings, Themes
+from app.impls import load_storage_config
 from app.logging import HealthCheckFilter
 from app.models import Feed, HealthCheck
 from app.rss import PrecisRSS
-from app.storage.engine import load_storage_config
+from app.settings import GlobalSettings, Themes
 
 JSON = "application/json"
